requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
motor==3.7.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
pytz==2025.1
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
Telethon==1.38.1
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
